refactor tests into separate file test_module_containers.py
